lab2 sol

+++ inc/string.h
@@ -20,6 +20,6 @@ void *	memmove(void *dst, const void *src, size_t len);
 int	memcmp(const void *s1, const void *s2, size_t len);
 void *	memfind(const void *s, int c, size_t len);
 
-long	strtol(const char *s, char **endptr, int base);
+long long	strtol(const char *s, char **endptr, int base);
 
 #endif /* not JOS_INC_STRING_H */
